Frame out new data structures, functions and methods needed to parse resources and the attribution object
